backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.7.0
spacy==3.8.1
nltk==3.9.1
scikit-learn==1.6.1
boto3==1.35.47
supabase==2.9.1
python-dotenv==1.0.1
httpx==0.28.1
wordfreq==3.0.1

